small fixes for inputhandler; base functionality for unit to iteract with world;

+++ src/main/java/org/game/towers/interfaces/IUnit.java
@@ -40,11 +40,11 @@ public interface IUnit {
 	public IUnit  setTileY(int tileY);
 	public int    getTileY();
 
-	public void  setX(int x);
-	public int    getX();
+	public void  	setX(double x);
+	public double	getX();
 
-	public void  setY(int y);
-	public int    getY();
+	public void  	setY(double y);
+	public double   getY();
 
     public boolean isConstruction();
 }
